# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oracle Clip Production Hub API",
    "description": "Deterministic media processing, opportunity generation, clip validation, and rendering service.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/lifecycle/usage/{target_tenant_id}": {
      "get": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Get Tenant Lifecycle Usage",
        "description": "Returns disk and quota telemetry for the tenant.",
        "operationId": "get_tenant_lifecycle_usage_v1_lifecycle_usage__target_tenant_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "target_tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Target Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lifecycle/cleanup": {
      "post": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Trigger Lifecycle Cleanup",
        "description": "Runs automated temporary file and failed render cleanup.",
        "operationId": "trigger_lifecycle_cleanup_v1_lifecycle_cleanup_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/config/summary": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Config Summary",
        "description": "Returns safe, sanitized configuration summary without credentials.",
        "operationId": "get_config_summary_v1_config_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/opportunities/generate": {
      "post": {
        "tags": [
          "opportunities"
        ],
        "summary": "Generate Opportunities",
        "description": "Phase 13.1: Deterministic Zero-LLM ContentOpportunity generation.",
        "operationId": "generate_opportunities_v1_opportunities_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateOpportunitiesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/clip-specs/validate": {
      "post": {
        "tags": [
          "clip_spec"
        ],
        "summary": "Validate Clip Specification",
        "description": "Phase 13.2: Structural ClipSpecification validation.",
        "operationId": "validate_clip_specification_v1_clip_specs_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateClipSpecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/render-jobs": {
      "post": {
        "tags": [
          "rendering"
        ],
        "summary": "Create Render Job",
        "description": "Submits and renders a ClipSpecification with strict tenant isolation and idempotency.",
        "operationId": "create_render_job_v1_render_jobs_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRenderJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/render-jobs/{job_id}": {
      "get": {
        "tags": [
          "rendering"
        ],
        "summary": "Get Render Job",
        "description": "Retrieves RenderJob state with tenant scoping.",
        "operationId": "get_render_job_v1_render_jobs__job_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/quality/check": {
      "post": {
        "tags": [
          "quality"
        ],
        "summary": "Check Asset Quality",
        "description": "Phase 13.4: Evaluates quality metrics on a rendered asset.",
        "operationId": "check_asset_quality_v1_quality_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderedAssetInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/guardian/evaluate": {
      "post": {
        "tags": [
          "quality"
        ],
        "summary": "Evaluate Guardian Policy",
        "description": "Phase 13.4: Evaluates Guardian policy for publication approval.",
        "operationId": "evaluate_guardian_policy_v1_guardian_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardianEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/webhooks/inbound": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Receive Inbound Webhook",
        "description": "Phase 13.7: Secure webhook receiver with HMAC-SHA256 signature verification and replay prevention.",
        "operationId": "receive_inbound_webhook_v1_webhooks_inbound_post",
        "parameters": [
          {
            "name": "X-Webhook-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Webhook-Signature"
            }
          },
          {
            "name": "X-Webhook-Timestamp",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Webhook-Timestamp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/checksums/verify": {
      "post": {
        "tags": [
          "checksums"
        ],
        "summary": "Verify Checksum Endpoint",
        "description": "Cryptographic SHA-256 Checksum Verification endpoint.",
        "operationId": "verify_checksum_endpoint_v1_checksums_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChecksumVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/orchestration/run": {
      "post": {
        "tags": [
          "orchestration"
        ],
        "summary": "Run Orchestration Pipeline",
        "description": "Executes full canonical media pipeline end-to-end with tenant authorization.",
        "operationId": "run_orchestration_pipeline_v1_orchestration_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunOrchestrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "GenerateOpportunitiesRequest": {
        "properties": {
          "source_media_id": {
            "type": "string",
            "title": "Source Media Id"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "scenes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SceneInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenes"
          },
          "max_opportunities": {
            "type": "integer",
            "title": "Max Opportunities",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "source_media_id",
          "duration_ms"
        ],
        "title": "GenerateOpportunitiesRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}